[TASK] Raise phpunit/phpunit compatibility to ^8.0 and ^9.0
