+++ tests/Packages/testbase/Tests/Functional/Controller/DeprecationControllerDisabledExceptionTest.php
@@ -31,7 +31,7 @@ class DeprecationControllerDisabledExceptionTest extends FunctionalTestCase
      */
     protected $deprecationController;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
