+++ tests/Packages/testbase/Tests/Functional/Controller/DeprecationControllerTypo3ExceptionTest.php
@@ -32,7 +32,7 @@ class DeprecationControllerTypo3ExceptionTest extends FunctionalTestCase
      */
     protected $deprecationController;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         if (!defined('ORIGINAL_ROOT')) {
             $this->markTestSkipped('Functional tests must be called through phpunit on CLI');
@@ -47,7 +47,7 @@ class DeprecationControllerTypo3ExceptionTest extends FunctionalTestCase
     public function someDeprecatedMethodThrowsDeprecationMessage()
     {
         $this->expectException(Exception::class);
-        $this->expectExceptionMessageRegExp('/^(PHP :|PHP User Warning:)/');
+        $this->expectExceptionMessageMatches('/^(PHP :|PHP User Warning:)/');
 
         parent::setUp();
