+++ tests/Packages/testbase/Tests/Unit/ViewHelpers/RenderChildrenViewHelperTest.php
@@ -23,7 +23,7 @@ class RenderChildrenViewHelperTest extends ViewHelperBaseTestcase
      */
     protected $viewHelper;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
         $this->viewHelper = $this->getMockBuilder(RenderChildrenViewHelper::class)
